Move sm/misc/macro.hs to common/macros.h
This file is not specific to the StorageManager so we can move it out of the
`sm` directory. As a side effect of my find-and-replace, a few newlines were
added to the end of some files. We should be doing this anyway so I left them
in this patch.

The motivation is that the common/heap_profiler.h/cc I'm writing uses these
macros and I don't want files in common/ to depend on files in sm/. We have
this problem with the existing common/threadpool.h depending on this macros.h
file as well.
