+++ tiledb/sm/cache/lru_cache.h
@@ -34,8 +34,8 @@
 #define TILEDB_LRU_CACHE_H
 
 #include "tiledb/common/logger.h"
+#include "tiledb/common/macros.h"
 #include "tiledb/common/status.h"
-#include "tiledb/sm/misc/macros.h"
 
 #include <list>
 #include <mutex>
