+++ tiledb/sm/filesystem/mem_filesystem.cc
@@ -36,7 +36,6 @@
 
 #include "tiledb/common/logger.h"
 #include "tiledb/sm/filesystem/mem_filesystem.h"
-#include "tiledb/sm/misc/macros.h"
 #include "tiledb/sm/misc/utils.h"
 
 using namespace tiledb::common;
