+++ tiledb/sm/filesystem/mem_filesystem.h
@@ -37,8 +37,8 @@
 #include <unordered_map>
 #include <vector>
 
+#include "tiledb/common/macros.h"
 #include "tiledb/common/status.h"
-#include "tiledb/sm/misc/macros.h"
 
 using namespace tiledb::common;
