+++ tiledb/sm/filesystem/vfs.h
@@ -39,6 +39,7 @@
 #include <string>
 #include <vector>
 
+#include "tiledb/common/macros.h"
 #include "tiledb/common/status.h"
 #include "tiledb/common/thread_pool.h"
 #include "tiledb/sm/buffer/buffer.h"
@@ -47,7 +48,6 @@
 #include "tiledb/sm/filesystem/filelock.h"
 #include "tiledb/sm/filesystem/mem_filesystem.h"
 #include "tiledb/sm/misc/cancelable_tasks.h"
-#include "tiledb/sm/misc/macros.h"
 #include "tiledb/sm/misc/uri.h"
 
 #ifdef _WIN32
