+++ tiledb/sm/query/validity_vector.h
@@ -36,8 +36,8 @@
 #include <vector>
 
 #include "tiledb/common/logger.h"
+#include "tiledb/common/macros.h"
 #include "tiledb/common/status.h"
-#include "tiledb/sm/misc/macros.h"
 
 using namespace tiledb::common;
